requirements.txt
beautifulsoup4==4.12.3
lxml==5.2.2
pandas==2.2.2
python-dotenv==1.0.1
requests==2.32.3
selenium==4.21.0
undetected-chromedriver==3.5.5
email-validator==2.1.1